- Updated the path for deleting images Unlink method. - Added Routes. - Registration Validations changed.

+++ app/Http/Controllers/ImageBankController.php
@@ -65,12 +65,12 @@ class ImageBankController extends Controller
             if (env('APP_ENV') == 'local') {
                 $path = public_path('storage/image-bank');
             } else {
-                $path = '/home/kohin837/public_html/preparemedicine.com/storage/image-bank';
+                $path = env('STORAGE_PATH').'/image-bank';
             }
 
             //now check directory
             if (!file_exists($path)) {
-                mkdir($path, 0777, true);
+                mkdir($path, 0775, true);
             }
 
             //now move upload image ok
@@ -161,16 +161,18 @@ class ImageBankController extends Controller
             if (env('APP_ENV') == 'local') {
                 $path = public_path('storage/image-bank');
             } else {
-                $path = '/home/kohin837/public_html/preparemedicine.com/storage/image-bank';
+                $path = env('STORAGE_PATH').'/image-bank';
             }
 
             //now check directory
             if (!file_exists($path)) {
-                mkdir($path, 0777, true);
+                mkdir($path, 0775, true);
             }
 
             /* Delete File before Uploading New */
-            unlink($path.'/'.$image->image);
+            if (file_exists($path.'/'.$image->image)) {
+                unlink($path.'/'.$image->image);
+            }
 
             //now move upload image ok
             $moved = $submitted_image->move($path, $img_uniqueName);
@@ -203,6 +205,10 @@ class ImageBankController extends Controller
     public function destroy($id)
     {
         $image = Ibank::find($id);
+
+        if (file_exists(public_path('storage/image-bank/'.$image->image))) {
+            unlink(public_path('storage/image-bank/'.$image->image));
+        }
         $image->delete();
 
         return redirect()->route('image-bank.index')
